For continuous tracking disable pebble timeout
During tracking, Pebble has a time out check for the next message to
occur.
For periodic tracking, the time for the next message is know so the time
out check by Pebble is valid.
For continuous tracking, the time for next next message is unknown
because it based on traveling a specified distance, and the time check
by pebble needs to be disabled.