Set tracking off when a trail is loaded
Previously, for Track On on the map bar, tracking would start
automatically when a trail map was selected and drawn. This caused
scaling problems for the compass arrow drawing as a map layer polygon.
It seems that the thread needs to end before drawing the arrow polygon
for the scaling to be correct.
Clearing tracking (Track Off on the map bar) avoids the scaling problem.
Also it seems better for me for tracking to be cleared when a trail is
selected. and drawn on the map.