Define Schema level for Settings
The Settings.nSchema property allows for changing default setting
properties for a new release of code.
Once in awhile it may be useful to change a default settings property.
The method for doing so allows for the default to be changed, but not to
override a setting a user may have set at a previous schema level. In
other words, a new release preserves previous settings unless a new
default specific to the release is specified.
In js/Model2.js GeoTrailSettings.LoadFromLocalStorage() and
GeoTrailSettings.SaveToLocalStorage() manage this.