Common Code for ios Distribution via HockeyApp.
Tag as release 1.1.022_20161215 at Github.
This is essentially save as v1.1.021_a, the release for Google Play
Store.
This is common code release for ios distribution via HockeyApp.
A branch to add recording a trail as you travel will added next.
Note that ios code is in the github geotrail-pgb repository and is a
copy of this code.