Inhibit loading a trail list after login for a recording a new trail.
Upon completion of login authentication, a list of trails used to be
always loaded. This worked fairly well, but could be confusing when
logging in to upload a new recorded trail because down loading a list of
trails interferred with uploading the new recorded trail. Therefore do
not download a list trails after login authentication if defining a new
trail trail name for upload.