Provide UI for setting Calorie Conversion Efficiency.
Show Stats for a Record Trail saves the stats to localStorage.
In Main Menu > Settings, based on the saved stats, the time, distance,
speed, kinetic calories, and calories burrned are shown along with
instructions for the user to look up the actual calories burned from an
independent, nutritional web site.  The conversion efficiency is updated
equal to  KineticEngery / ActualCalories.  User only needs to the enter
the ActualCalories; the efficiency is calculated and save.