Use SSL for transport.
Google Developer agreement was updated to provide secure transport of
personal data. Since the Facebook user id is a public, personally
identifier, https protocol should be used with the web server for the
geotrail RESTful api. I so changed to use https for getting leaflet
images.

Measure height dynanically for selectGeoTrail droplist so that it shows
better on iPhone.

Adjust CSS for better appearance on Pixel.