Fix for Pebble plugin.
The Pebble plugin was not working for Pixel. The problem was due to the
plugin not being initialized properly, probably a timing problem when
creating the wigo_ws_PebbleAdapter. Now the adapter tries to initialize
again when starting Pebble or sending text to Pebble if it was not
successfully initialized when the adapter is created. The initialization
involves attaching event handler to the document for events received
from Pebble.