Improve defining trail name for uploading and saving offline.
From Online View, Save Offline button when no trail is selected saves
the area displayed on the screen. A single point trail at the center of
the screen is saved offline.
Made defining a trail name less confusing showing only the bar with its
action buttons when the userr needs to enter a trail name. The map bar
at bottom of screen and the bar for selecting a trail from a droplist is
hidden.