Ensure click for geolocation is reset when app loads.
Previous change did this when app version changed. Decided to do
whenever app is reloaded after it has been closed.
Also filters before to updating distance interval for an alert due to a
Settings change to avoid problem of current distance being to long
because filtering has not been done.