Add back button to Close Dialog Bar for Help Guide
For the Help Guide, a back button is shown on the Close Dialog Bar when
a link to a reference has been taken. The back button (<---) returns to
the previous point, the scrollTop of the document, when the link was
taken. A stack of returns is kept. The stack is popped after a return.
If the stack is empty, the back button is not shown.

Only the Help Guide div show a back button, the other help divs do not.