Saving RecordStatsHistory at web server.
Upload RecordStatsHistory additions when switching to Stats  History
view.
On log out, save RecordStats objs that have not been uploaded.
When signing in, get the stats history for the user from the web server.
This has been tested some and seems to be working, but more testing is
needed and some changes for editing or adding a Record Stats obj from
Stats History View.