Add Walking Map View
Walking Map is now the initial view when app loads. A user can easily start recording a walk. This is similar to the the record droplist in the Online View (now called Trail Maps), but has a much simpler UI because there are only two button, Start/End and Pause/Resume, to control recording.